requirements.txt
fastapi
uvicorn
celery
redis
pymongo
motor
paho-mqtt
supabase
firebase-admin
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
pydantic[email]
python-dotenv
httpx
loguru
